Refactoring (#5)
* Use subdir

* Refactor notification

* Split and rename server part

* Split and rename client part

Move credentials logic inside

* Rename entrance files
